refactor(resiliency): split resilient client internals
